Move tasks from pkg/installer to pkg/tasks
