+++ pkg/tasks/static_wokers.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package installation
+package tasks
 
 import (
 	kubeoneapi "github.com/kubermatic/kubeone/pkg/apis/kubeone"
